refactor: apply some discussion about type, contant
