+++ apps/web/src/api-schema/identity/role/constant.ts
@@ -1,0 +1,5 @@
+export const ROLE_TYPE = {
+    SYSTEM: 'SYSTEM',
+    PROJECT: 'PROJECT',
+    DOMAIN: 'DOMAIN',
+} as const;
