+++ apps/web/src/api-schema/identity/user/constant.ts
@@ -1,0 +1,4 @@
+export const USER_TYPE = {
+    API_USER: 'API_USER',
+    USER: 'USER',
+} as const;
