+++ apps/web/src/services/administration/components/UserNotifications.vue
@@ -84,7 +84,7 @@ import { i18n } from '@/translations';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
-import { utcToTimezoneFormatter } from '@/services/administration/helpers/user-notification-timezone-helpers';
+import { utcToTimezoneFormatter } from '@/services/administration/helpers/user-notification-timezone-helper';
 import type { ChannelItem, ProtocolItem } from '@/services/administration/types/user-notification-type';
 import { MY_PAGE_ROUTE } from '@/services/my-page/route-config';
