+++ apps/web/src/services/administration/constants/policy-constant.ts
@@ -1,0 +1,5 @@
+export const POLICY_TYPE = {
+    MANAGED: 'MANAGED',
+    CUSTOM: 'CUSTOM',
+    ALL: 'ALL',
+} as const;
