+++ apps/web/src/services/administration/helpers/policy-helper.ts
@@ -2,9 +2,10 @@ import { iso8601Formatter } from '@cloudforet/core-lib';
 import type { KeyDataType, ValueHandler, ValueItem } from '@cloudforet/core-lib/component-util/query-search/type';
 
 import type { TimeStamp } from '@/api-schema/common/model';
-import { POLICY_TYPE } from '@/api-schema/identity/policy/model';
-import type { PolicyModel, PolicyType } from '@/api-schema/identity/policy/model';
+import type { PolicyModel } from '@/api-schema/identity/policy/model';
 
+import { POLICY_TYPE } from '@/services/administration/constants/policy-constant';
+import type { PolicyType } from '@/services/administration/types/policy-type';
 
 
 export const policyTypeBadgeColorFormatter = (policyType: PolicyType) => {
